Add the possibility to parametrize what is tested

+++ perfs/run.sh
@@ -24,17 +24,24 @@ do
     sleep 1
 done
 
-export base_urls="http://geoserver:8080/OSM/ows|GeoServer,http://mapserver/|MapServer,http://qgis2/|QGIS2,http://qgis3/|QGIS3"
+export base_urls="${BASE_URLS:-http://geoserver:8080/OSM/ows|GeoServer,http://mapserver/|MapServer,http://qgis2/|QGIS2,http://qgis3/|QGIS3}"
+echo "base_urls=$base_urls"
 
-echo "warmup round (results are trashed)"
-export nb_users=1
-export time=30
-gatling.sh -sf $GATLING_HOME/user-files/simulations -s com.camptocamp.Test
+warmup=${WARMUP:-1}
 
-rm -r $GATLING_HOME/results/*
+if [ $warmup == '1' ]
+then
+    echo "warmup round (results are trashed)"
+    export nb_users=1
+    export time=30
+    gatling.sh -sf $GATLING_HOME/user-files/simulations -s com.camptocamp.Test
 
-export time=120  # time to keep measing for a given server and a given zoom level
-for nb_users in 1 2 5 10 20 40
+    rm -r $GATLING_HOME/results/*
+fi
+
+export time=${TIME_PER_TEST:-120}  # time to keep measing for a given server and a given zoom level
+users=${USERS:-1 2 5 10 20 40}
+for nb_users in $users
 do
     echo "Measuring with $nb_users users in //"
     export nb_users=$nb_users
